Added army units behavior. Also a lot of small fixes.
Added the behaviors for the tank going into siege mode, the marine using
a stim-pack and the healer healing a valid target.

Notice: There will now be unCreated units within the needToAdd list.

Added function for getting every unit of certain type. Also changed old
implementation of this functionality to the newly added one. A more
stable unit creation environment can be seen.
Added a new class for commanding units which needs implementation.
Changed ->gather to ->rightClick for the OnFrame() idle workers
behavior.
Changed the logic for possitioning the factories by a slight amount.
They now correctly build away from the supply depots.

Made a botch to make sure that, in step 2, when there is a need to add
another unit and the supply is all used upp, create more supply. The
unit assigned to this task will be given the HiredID of 26.

Made a change in step3() so that the algorithm finds the research
facility capable of researching the tech the fastest.