Problems fixed, making final changes to the progression through step 1 through 3. Started on solution for step 4.
Still a problem where the unit count for a given unit is off by 1. If
there is 5 on the board and 4 training, the output will be 10 until
there are no units being trained at which point it will drop to 9 to
then increase to 10 with the new unit being trained.